FIX : validator 폴도 구조 및 chat, user DTO 작성
