+++ src/dtos/chat.dto.ts
@@ -1,0 +1,7 @@
+export interface RequestCreateChatRoomDTO{
+    u_id : number,    
+};
+
+export interface RequestCreateMessageDTO{
+    r_id : number,
+};
