+++ src/dtos/user.dto.ts
@@ -1,0 +1,5 @@
+export interface RequestUserDto{
+    u_email : string,
+    u_password : string,
+    u_nickname : string,
+}
